Refactor market, orderbook, pool, and admin domains into model/router/service layers
Market (#41):
- Create models/market.py (KLINE_INTERVALS)
- Create services/market.py (market data, kline generation with forward-fill)
- Rewrite routers/market.py as thin router (4 endpoints)

Orderbook (#42):
- Create models/orderbook.py (placeholder for future migration)
- Create services/orderbook.py (order queries, placement, cancellation)
- Rewrite routers/orderbook.py as thin router (7 endpoints)

Pool (#43):
- Create models/pool.py (PeriodKind, symbol parsing helpers)
- Create services/pool.py (pool queries, swap, liquidity, charts — largest service)
- Rewrite routers/pool.py as thin router (12 endpoints, was 867 lines)

Admin (#44):
- Create models/admin.py (placeholder for future migration)
- Create services/admin.py (symbol CRUD, pool creation, audit log queries)
- Rewrite routers/admin.py as thin router with @audit_logged decorator

All 45 endpoints verified working. No behavior change.

Closes #41
Closes #42
Closes #43
Closes #44
