+++ backend/models/admin.py
@@ -1,0 +1,6 @@
+"""
+Admin domain models.
+
+Request models (CreateSymbolRequest, CreatePoolRequest) are still
+in models/requests.py and will be migrated here during cleanup (#45).
+"""
