+++ backend/models/orderbook.py
@@ -1,0 +1,6 @@
+"""
+Orderbook domain models — CLOB orders and trades.
+
+Request models (PlaceOrderRequest) are still in models/requests.py
+and will be migrated here during cleanup (#45).
+"""
